Bug Fix
We were setting takingAdvanced command false twice. Once when
constructing the advanced command string from the advanced command
buffer, and again, in the keypressed handler.  No need to do it twice,
and I think the right place to do it is when the advanced command is
proccessed.  That way, no matter how we decide the advanced command is
completed, we get consistent behavior.